Reworked 'trivia admin cancel' command
The 'trivia admin cancel' command has been reworked. When typed
mid-round, the round results will be displayed before the game is
cancelled. This also fixes rounds not cancelling when the command is
typed between rounds.

Round timers are now set under game[id] so it is easier to modify round
timing in the future.

See issues #44 and #43