Added generic fix for frozen trivia games
This is a generic fix for trivia games getting stuck. If a user attempts
to start a game, this will check if an existing game is stuck.

If a game is stuck, an error while display in console. However, the user
will be able to continue playing, rather than no-longer being able to
play in that channel.

Note that this is only a generic fix until the root cause of games
freezing can be found.