Add answer check
This checks to make sure that game timers always match their corresponding answers. If there is a mismatch, an error is outputted and the answer does not display.

This is a fix for the effects of issue #65, however it is not a real solution and will still output a console error.