Add config option: rounds-end-after
This is a new config option that changes how many inactive rounds are allowed without the game ending.

By default this is set to 2, meaning that on the second round of inactivity, the game will end. Setting this to 1 causes the bot to behave like previous versions, ending the game after a single round of inactivity.

To compensate this change, the round timer has been decreased from 25 seconds to 20 seconds. This shortens the wait time for non-admin users that are waiting for the round to end.