Game now exports on client error
The bot now automatically uses the maintenance export to retain game data if a client error occurs. This will reduce instances of games stopping suddenly due to an error.

This has been tested to ensure that the bot will not get stuck in a loop if importing/resuming the game directly causes another client error.