Fix scores potentially resetting to N/A
Scores now have a failsafe to prevent them from resetting and displaying as "N/A" if the difficulty value has not been specified properly.

This has several causes, some of which are not fully known and still need to be investigated, however this will aid in this process and it prevents the problem from completely breaking the game.

See issue #132