Refactor config data passthrough
Config data is now passed to scripts using a "configData" object, which contains the actual config in the "config" key. This is to make it easier to pass more detailed information that is separate from the actual config data (such as the config file path).

As a result of this, the bot no-longer needs to strip the configFile key every time it writes a config value.