Removed showing discord.js version
Just an idea, as you are not using always the latest version of discord.js, showing what old version you are using is giving hackers option to check what vulnerabilities are in this bot and they may use them. Also I think showing the exact version is not helping anyone, now it's on your consideration.