Migrate doGame functionality
Migrate the functionality of doGame to event emitters for the Game and GameHandler objects. Repace all uses of doGame with a new Game call. The goal is to consolidate most or all Discord-specific functionality to event handlers.

Use channel cache finding to retain the channel data for the game.

As a temporary solution, also use the channel cache finder for doAnswerReveal.